Rename MessageCenter to MessageRouter
